[reduce_raw.sh] first try to call match_filter script

+++ scripts/reduce_raw.sh
@@ -10,8 +10,14 @@
 HOME_DIR=${HOME_DIR:-.}
 RES_DIR=${RES_DIR:-results}
 CFG_DIR=${CFG_DIR:-config}
+BIN_DIR=${BIN_DIR:-scripts}
 
-#set -e
+# flags for match filter and reference reduction
+#
+MF=
+REFERENCE=
+
+set -e
 
 
 if [ $# -ne 1 ];then
@@ -21,13 +27,6 @@ fi
 
 inconf=$1
 source $inconf
-
-
-#echo;echo "  + file list:" $LIST_FILE
-#echo "  + data directory:" $ORAC_DATA_IN
-#echo "  + results directory:" $ORAC_DATA_OUT
-#echo "  + reduction recipe:" $RECIPE
-#echo "  + parameters:" $PARAMS;echo
 
 
 case $FRQ in
@@ -46,15 +45,6 @@ source $inconf
 ORAC_DATA_OUT=${RES_DIR}/${ORAC_DATA_OUT}
 LIST_FILE=${HOME_DIR}/${LIST_FILE}
 
-
-echo;echo "  + file list:" $LIST_FILE
-echo "  + data directory:" $ORAC_DATA_IN
-echo "  + results directory:" $ORAC_DATA_OUT
-echo "  + reduction recipe:" $RECIPE
-echo "  + parameters:" $PARAMS
-echo
-
-
 export ORAC_DATA_IN
 export ORAC_DATA_OUT
 
@@ -65,32 +55,79 @@ fi
 
 LOGFILE=${ORAC_DATA_OUT}/oracdr.log
 
-oracdr  -files $LIST_FILE  -loop file  -nodisplay  \
-	-log sf  -verbose $RECIPE  -recpars $PARAMS \
-    |tee -a $LOGFILE
+RESULT_FILE=${ORAC_DATA_OUT}/${OUTPUT_FILE}
 
 
+echo
+echo "  + file list:" $LIST_FILE |tee -a $LOGFILE
+echo "  + data directory:" $ORAC_DATA_IN |tee -a $LOGFILE
+echo "  + results directory:" $ORAC_DATA_OUT  |tee -a $LOGFILE
+echo "  + reduction recipe:" $RECIPE  |tee -a $LOGFILE
+echo "  + parameters:" $PARAMS  |tee -a $LOGFILE
+echo
 
-if [[ -n ${ORAC_DATA_OUT}/${REDUCED_FILE} && \
-	  -n ${ORAC_DATA_OUT}/${OUTPUT_FILE} ]]
+
+# if we have a reference reduced file, for future processing, we do not
+# process the data again and copy the reference file as the reduced file
+#
+# this will be used to apply later a match filter with another beam
+#
+if [[ -z $REFERENCE ]]
 then
-    echo "Copying to $OUTPUT_FILE"
-    mv ${ORAC_DATA_OUT}/${REDUCED_FILE} ${ORAC_DATA_OUT}/${OUTPUT_FILE} |
-       tee -a $LOGFILE
+
+    oracdr -files $LIST_FILE -loop file -nodisplay -log sf -verbose $RECIPE \
+	   -recpars $PARAMS |tee -a $LOGFILE
+
+    
+    if [[ -n ${ORAC_DATA_OUT}/${REDUCED_FILE} && \
+	      -n ${RESULT_FILE} ]]
+    then
+	echo "Copying to $OUTPUT_FILE"
+	mv ${ORAC_DATA_OUT}/${REDUCED_FILE} ${RESULT_FILE} \
+	    | tee -a $LOGFILE
+    fi
+
+
+    reducbl=${ORAC_DATA_OUT}/reduced_blocks
+    if [[ ! -d  $reducbl ]]
+    then
+	mkdir ${reducbl}
+    fi
+
+    echo;echo "   moving data to reduced_blocks..."
+    cd ${ORAC_DATA_OUT}
+    rm *256.png *64.png
+    mv s*.sdf s*.png  reduced_blocks/
+    mv log* reduced_blocks/
+    mv gs*.sdf gs*.png  gs*.FIT reduced_blocks/
+    echo "   ... done";echo
+
+else
+    if [[ -f $REFERENCE ]]
+    then
+	echo;echo " Copying $REFERENCE to ${RESULT_FILE} ..."|tee -a $LOGFILE
+	cp $REFERENCE ${RESULT_FILE}
+    else
+	echo;echo " ERROR: $REFERENCE could not be found"
+	exit 2
+    fi
 fi
 
 
-reducbl=${ORAC_DATA_OUT}/reduced_blocks
-if [[ ! -d  $reducbl ]]
+if [[ $MF == y ]]
 then
-    mkdir ${reducbl}
+    echo;echo;echo " ... applying match filter";echo |tee -a $LOGFILE
+    
+    OLDFILE=${RESULT_FILE/.sdf/-original.sdf}
+    
+    ${BIN_DIR}/make_match_filter.sh ${RESULT_FILE} ${RESULT_FILE} \
+	      ${BEAM_FILE} |tee -a $LOGFILE
+
+    if [[ -n ${CALIB_FILE} ]]
+    then
+	CALIB_REDUC=${ORAC_DATA_OUT}/${CALIB_FILE}
+	${BIN_DIR}/make_match_filter.sh ${CALIB_REDUC} ${CALIB_REDUC} \
+		  ${BEAM_FILE} |tee -a $LOGFILE
+    fi
+    
 fi
-
-echo;echo "   moving data to reduced_blocs..."
-cd ${ORAC_DATA_OUT}
-rm *256.png *64.png
-mv s*.sdf s*.png  reduced_blocks/
-mv log* reduced_blocks/
-mv gs*.sdf gs*.png  gs*.FIT reduced_blocks/
-echo "   ... done";echo
-
